Create core.magics.osm according to new API.
